requirements.txt
flask==3.1.2
requests==2.32.4
python-dotenv==1.2.1